Terrible, awful, dirty hack to fix admin map
It's actually fixed now! For real this time!

This code is terrible, but it's all I could really think of doing that
wasn't super convoluted or messy. VGUI has forced my hand with this
hack, but there's not much better I can do. And hey, at least it doesn't
cause a memory leak or anything.